Rector: updated syntax to php 81

+++ app/code/community/Aoe/Scheduler/Model/ScheduleManager.php
@@ -27,8 +27,8 @@ class Aoe_Scheduler_Model_ScheduleManager
 
         Mage::getSingleton('cron/schedule')->getResource()->beginTransaction(TRUE);
         try {
-            $seenJobs = array();
-            foreach ($schedules as $key => $schedule) {
+            $seenJobs = [];
+            foreach ($schedules as $schedule) {
                 /* @var Aoe_Scheduler_Model_Schedule $schedule */
                 if (isset($seenJobs[$schedule->getJobCode()])) {
                     $schedule
@@ -58,8 +58,6 @@ class Aoe_Scheduler_Model_ScheduleManager
     /**
      * Get pending schedules
      *
-     * @param array $whitelist
-     * @param array $blacklist
      *
      * @return Mage_Cron_Model_Resource_Schedule_Collection
      */
@@ -70,13 +68,13 @@ class Aoe_Scheduler_Model_ScheduleManager
             ->addFieldToFilter('scheduled_at', ['lt' => date('Y-m-d H:i:s', time())])
             ->addOrder('scheduled_at', 'ASC');
 
-        $whitelist = array_filter(array_map('trim', $whitelist));
-        if (!empty($whitelist)) {
+        $whitelist = array_filter(array_map(trim(...), $whitelist));
+        if ($whitelist !== []) {
             $pendingSchedules->addFieldToFilter('job_code', ['in' => $whitelist]);
         }
 
-        $blacklist = array_filter(array_map('trim', $blacklist));
-        if (!empty($blacklist)) {
+        $blacklist = array_filter(array_map(trim(...), $blacklist));
+        if ($blacklist !== []) {
             $pendingSchedules->addFieldToFilter('job_code', ['nin' => $blacklist]);
         }
 
@@ -98,7 +96,7 @@ class Aoe_Scheduler_Model_ScheduleManager
             $ts = date('Y-m-d H:i:00', time());
             $schedule = Mage::getModel('cron/schedule'); /* @var $schedule Aoe_Scheduler_Model_Schedule */
             $schedule
-                ->setScheduledReason($reason ? $reason : Aoe_Scheduler_Model_Schedule::REASON_ALWAYS)
+                ->setScheduledReason($reason ?: Aoe_Scheduler_Model_Schedule::REASON_ALWAYS)
                 ->setJobCode($jobCode)
                 ->setStatus(Aoe_Scheduler_Model_Schedule::STATUS_RUNNING)
                 ->setCreatedAt($ts)
@@ -131,7 +129,7 @@ class Aoe_Scheduler_Model_ScheduleManager
 			HAVING qty > 1;
 		");
         foreach ($results as $row) {
-            $ids = explode(',', $row['ids']);
+            $ids = explode(',', (string) $row['ids']);
             $removeIds = array_slice($ids, 1);
             foreach ($removeIds as $id) {
                 Mage::getModel('cron/schedule')->load($id)->delete();
@@ -199,16 +197,16 @@ class Aoe_Scheduler_Model_ScheduleManager
         /* @var $pendingSchedules Mage_Cron_Model_Resource_Schedule_Collection */
         $pendingSchedules = Mage::getModel('cron/schedule')->getCollection()
             ->addFieldToFilter('status', Aoe_Scheduler_Model_Schedule::STATUS_PENDING)
-            ->addFieldToFilter('scheduled_at', array('gt' => date('Y-m-d H:i:s', time())))
+            ->addFieldToFilter('scheduled_at', ['gt' => date('Y-m-d H:i:s', time())])
             ->addOrder('scheduled_at', 'ASC');
         if (!empty($jobCode)) {
             $pendingSchedules->addFieldToFilter('job_code', $jobCode);
         }
-        foreach ($pendingSchedules as $key => $schedule) {
+        foreach ($pendingSchedules as $schedule) {
             /* @var Aoe_Scheduler_Model_Schedule $schedule */
             $schedule->delete();
         }
-        Mage::app()->saveCache(0, Mage_Cron_Model_Observer::CACHE_KEY_LAST_SCHEDULE_GENERATE_AT, array('crontab'), null);
+        Mage::app()->saveCache(0, Mage_Cron_Model_Observer::CACHE_KEY_LAST_SCHEDULE_GENERATE_AT, ['crontab'], null);
         return $this;
     }
 
@@ -221,17 +219,16 @@ class Aoe_Scheduler_Model_ScheduleManager
     {
         /* @var $schedules Mage_Cron_Model_Resource_Schedule_Collection */
         $schedules = Mage::getModel('cron/schedule')->getCollection();
-        foreach ($schedules as $key => $schedule) { /* @var Aoe_Scheduler_Model_Schedule $schedule */
+        foreach ($schedules as $schedule) { /* @var Aoe_Scheduler_Model_Schedule $schedule */
             $schedule->delete();
         }
-        Mage::app()->saveCache(0, Mage_Cron_Model_Observer::CACHE_KEY_LAST_SCHEDULE_GENERATE_AT, array('crontab'), null);
+        Mage::app()->saveCache(0, Mage_Cron_Model_Observer::CACHE_KEY_LAST_SCHEDULE_GENERATE_AT, ['crontab'], null);
         return $this;
     }
 
     /**
      * Generate jobs for config information
      *
-     * @param Aoe_Scheduler_Model_Job $job
      *
      * @return $this
      */
@@ -241,8 +238,8 @@ class Aoe_Scheduler_Model_ScheduleManager
             return $this;
         }
 
-        $exists = array();
-        foreach ($this->getPendingSchedules(array($job->getJobCode()), array()) as $schedule) {
+        $exists = [];
+        foreach ($this->getPendingSchedules([$job->getJobCode()], []) as $schedule) {
             /* @var Aoe_Scheduler_Model_Schedule $schedule */
             $exists[$schedule->getJobCode() . '/' . $schedule->getScheduledAt()] = 1;
         }
@@ -257,7 +254,7 @@ class Aoe_Scheduler_Model_ScheduleManager
 
         for ($time = $now + 60; $time < $timeAhead; $time += 60) {
             $ts = date('Y-m-d H:i:00', $time);
-            if (!empty($exists[$job->getJobCode().'/'.$ts])) {
+            if (isset($exists[$job->getJobCode().'/'.$ts]) && $exists[$job->getJobCode().'/'.$ts] !== 0) {
                 // already scheduled
                 continue;
             }
@@ -288,13 +285,13 @@ class Aoe_Scheduler_Model_ScheduleManager
         $startTime = microtime(true);
 
         $history = Mage::getModel('cron/schedule')->getCollection()
-            ->addFieldToFilter('status', array('nin' => array(
+            ->addFieldToFilter('status', ['nin' => [
                 Aoe_Scheduler_Model_Schedule::STATUS_PENDING,
                 Aoe_Scheduler_Model_Schedule::STATUS_RUNNING
-            )))
+            ]])
             ->load();
 
-        $historyLifetimes = array(
+        $historyLifetimes = [
             Aoe_Scheduler_Model_Schedule::STATUS_KILLED =>          Mage::getStoreConfig(Mage_Cron_Model_Observer::XML_PATH_HISTORY_SUCCESS)*60,
             Aoe_Scheduler_Model_Schedule::STATUS_DISAPPEARED =>     Mage::getStoreConfig(Mage_Cron_Model_Observer::XML_PATH_HISTORY_FAILURE)*60,
             Aoe_Scheduler_Model_Schedule::STATUS_DIDNTDOANYTHING => Mage::getStoreConfig(Mage_Cron_Model_Observer::XML_PATH_HISTORY_SUCCESS)*60,
@@ -305,19 +302,17 @@ class Aoe_Scheduler_Model_ScheduleManager
             Aoe_Scheduler_Model_Schedule::STATUS_ERROR =>           Mage::getStoreConfig(Mage_Cron_Model_Observer::XML_PATH_HISTORY_FAILURE)*60,
             Aoe_Scheduler_Model_Schedule::STATUS_DIED =>            Mage::getStoreConfig(Mage_Cron_Model_Observer::XML_PATH_HISTORY_FAILURE)*60,
             Aoe_Scheduler_Model_Schedule::STATUS_SKIP_OTHERJOBRUNNING => Mage::getStoreConfig(Mage_Cron_Model_Observer::XML_PATH_HISTORY_FAILURE)*60,
-        );
+        ];
 
         $now = time();
         foreach ($history->getIterator() as $record) { /* @var $record Aoe_Scheduler_Model_Schedule */
-            if (isset($historyLifetimes[$record->getStatus()])) {
-                if (strtotime($record->getExecutedAt()) < $now - $historyLifetimes[$record->getStatus()]) {
-                    $record->delete();
-                }
+            if (isset($historyLifetimes[$record->getStatus()]) && strtotime($record->getExecutedAt()) < $now - $historyLifetimes[$record->getStatus()]) {
+                $record->delete();
             }
         }
 
         // save time history cleanup was ran with no expiration
-        Mage::app()->saveCache(time(), Mage_Cron_Model_Observer::CACHE_KEY_LAST_HISTORY_CLEANUP_AT, array('crontab'), null);
+        Mage::app()->saveCache(time(), Mage_Cron_Model_Observer::CACHE_KEY_LAST_HISTORY_CLEANUP_AT, ['crontab'], null);
 
 
         // delete successful tasks (beyond the configured max number of tasks to keep)
@@ -325,15 +320,15 @@ class Aoe_Scheduler_Model_ScheduleManager
         if ($maxNo) {
             $history = Mage::getModel('cron/schedule')->getCollection()
                 ->addFieldToFilter(
-                    array('status'),
-                    array(
-                        array('eq' => Aoe_Scheduler_Model_Schedule::STATUS_SUCCESS),
-                        array('eq' => Aoe_Scheduler_Model_Schedule::STATUS_REPEAT)
-                    )
+                    ['status'],
+                    [
+                        ['eq' => Aoe_Scheduler_Model_Schedule::STATUS_SUCCESS],
+                        ['eq' => Aoe_Scheduler_Model_Schedule::STATUS_REPEAT]
+                    ]
                 )
                 ->setOrder('finished_at', 'desc')
                 ->load();
-            $counter = array();
+            $counter = [];
             foreach ($history->getIterator() as $record) { /* @var $record Aoe_Scheduler_Model_Schedule */
                 $jobCode = $record->getJobCode();
                 if (!isset($counter[$jobCode])) {
@@ -363,7 +358,7 @@ class Aoe_Scheduler_Model_ScheduleManager
         $lastRuns = explode(',', $lastRuns);
         $lastRuns[] = time();
         $lastRuns = array_slice($lastRuns, -100);
-        Mage::app()->saveCache(implode(',', $lastRuns), self::CACHE_KEY_SCHEDULER_LASTRUNS, array('crontab'), null);
+        Mage::app()->saveCache(implode(',', $lastRuns), self::CACHE_KEY_SCHEDULER_LASTRUNS, ['crontab'], null);
     }
 
     /**
@@ -379,18 +374,18 @@ class Aoe_Scheduler_Model_ScheduleManager
             // not enough data points
             return false;
         }
-        $gaps = array();
-        foreach ($lastRuns as $index => $run) {
+        $gaps = [];
+        foreach (array_keys($lastRuns) as $index) {
             if ($index > 0) {
                 $gaps[$index] = intval($lastRuns[$index]) - intval($lastRuns[$index-1]);
             }
         }
-        return array(
+        return [
             'average' => round((array_sum($gaps) / count($gaps)) / 60, 2),
             'max' => round(max($gaps) / 60, 2),
             'min' => round(min($gaps) / 60, 2),
             'count' => count($gaps),
             'last' => end($lastRuns)
-        );
+        ];
     }
 }
